+++ functions/raddStoreRegistryLambda/src/app/csvUtils.js
@@ -1,41 +1,4 @@
-const validFieldValue = [
-  'locationId',
-  'partnerId',
-  'externalCodes',
-  'description',
-  'city',
-  'address',
-  'normalizedAddress',
-  'province',
-  'zipCode',
-  'phoneNumbers',
-  'monday',
-  'tuesday',
-  'wednesday',
-  'thursday',
-  'friday',
-  'saturday',
-  'sunday',
-  'latitude',
-  'longitude',
-  'rawOpeningHours',
-  'email',
-  'website',
-  'appointmentRequired',
-];
-
-const wrongAddressesConfig = [
-  { header: 'locationId', field: 'locationId' },
-  { header: 'partnerId', field: 'partnerId' },
-  { header: 'descrizione', field: 'description' },
-  { header: 'indirizzo_originale', field: 'cafAddress' },
-  { header: 'indirizzo_normalizzato', field: 'normalizedAddress' },
-  { header: 'score_AWS', field: 'biasPoint' },
-];
-
-const wrongAddressesCsvHeader = wrongAddressesConfig
-  .map((config) => config.header)
-  .join(';');
+const { validFieldValue } = require('../data/csvData');
 
 function validateCsvConfiguration(csvConfiguration) {
   console.log('Validating configuration');
@@ -68,29 +31,6 @@ function createCSVContent(configs, data) {
   return csvContent;
 }
 
-const getOpeningTimeByDay = (openingTimeObj) => {
-  const times = new Array(7).fill(null);
-
-  const dayMapping = {
-    lun: 0,
-    mar: 1,
-    mer: 2,
-    gio: 3,
-    ven: 4,
-    sab: 5,
-    dom: 6,
-  };
-
-  for (const [dayKey, hours] of Object.entries(openingTimeObj)) {
-    const dayIndex = dayMapping[dayKey.toLowerCase()];
-    if (dayIndex !== undefined) {
-      times[dayIndex] = hours;
-    }
-  }
-
-  return times;
-};
-
 function sanitizeCSVField(field) {
   if (field == null) return '';
 
@@ -109,43 +49,8 @@ function sanitizeCSVField(field) {
   return `"${escapedField}"`;
 }
 
-/**
- * Check if the AWS address is valid based on scores and thresholds
- *
- * @param {Object} scores - AWS address scores object (subRegion, locality, postalCode, addressNumber, overall)
- * @returns {boolean} True if address is valid, false otherwise
- */
-function isAWSAddressValid(scores) {
-  if (!scores) return false;
-
-  const requiredScores = [
-    'subRegion',
-    'locality',
-    'postalCode',
-    'addressNumber',
-  ];
-
-  const thresholds = {
-    subRegion: parseFloat(process.env.SUBREGION_THRESHOLD),
-    locality: parseFloat(process.env.LOCALITY_THRESHOLD),
-    postalCode: parseFloat(process.env.POSTAL_CODE_THRESHOLD),
-    addressNumber: parseFloat(process.env.ADDRESS_NUMBER_THRESHOLD),
-    overall: parseFloat(process.env.OVERALL_THRESHOLD),
-  };
-
-  return (
-    requiredScores.every(
-      (field) => scores[field] != null && scores[field] >= thresholds[field]
-    ) && scores.overall >= thresholds.overall
-  );
-}
-
 module.exports = {
   validateCsvConfiguration,
   createCSVContent,
-  getOpeningTimeByDay,
   sanitizeCSVField,
-  isAWSAddressValid,
-  wrongAddressesCsvHeader,
-  wrongAddressesConfig,
 };
