+++ functions/raddStoreRegistryLambda/src/app/eventHandler.js
@@ -4,15 +4,20 @@ const csvUtils = require('./csvUtils');
 const apiClient = require('./raddClient');
 const utils = require('./utils');
 const storeLocatorCsvEntity = require('./StoreLocatorCsvEntity');
+const {
+  wrongAddressesCsvHeader,
+  wrongAddressesConfig,
+} = require('../data/csvData');
 
 exports.handleEvent = async () => {
   console.log('Handler invoked');
-  validateEnvironmentVariables();
+  ssmUtils.validateEnvironmentVariables();
 
   let forceGenerate = false;
   let sendToWebLanding = false;
 
   const malformedAddressS3Key = `${process.env.BFF_BUCKET_PREFIX}/malformed_addresses.csv`;
+  const wrongPostalCodesS3Key = `${process.env.BFF_BUCKET_PREFIX}/wrong_postal_codes.csv`;
   const generationConfig = await ssmUtils.retrieveGenerationConfigParameter();
 
   if (generationConfig) {
@@ -48,7 +53,8 @@ exports.handleEvent = async () => {
     .map((conf) => conf.header)
     .join(';');
   let csvContent = csvHeader;
-  let wrongAddressesCsvContent = csvUtils.wrongAddressesCsvHeader;
+  let wrongAddressesCsvContent = wrongAddressesCsvHeader;
+  let wrongPostalCodesCsvContent = wrongAddressesCsvHeader;
 
   let lastKey = null;
 
@@ -64,9 +70,18 @@ exports.handleEvent = async () => {
     );
 
     for (let record of records) {
+      // Append to wrong addresses CSV file if the address is not valid
       if (!record.isRecordValid) {
         wrongAddressesCsvContent += csvUtils.createCSVContent(
-          csvUtils.wrongAddressesConfig,
+          wrongAddressesConfig,
+          [record.record]
+        );
+      }
+
+      // Append to wrong postal codes CSV file if the postal code is not valid
+      if (!record.isCAPValid) {
+        wrongPostalCodesCsvContent += csvUtils.createCSVContent(
+          wrongAddressesConfig,
           [record.record]
         );
       }
@@ -87,6 +102,13 @@ exports.handleEvent = async () => {
     wrongAddressesCsvContent
   );
 
+  // Upload wrong postal codes CSV file
+  await s3Utils.uploadVersionedFile(
+    false,
+    wrongPostalCodesS3Key,
+    wrongPostalCodesCsvContent
+  );
+
   // Upload store locator CSV file
   await s3Utils.uploadVersionedFile(
     sendToWebLanding,
@@ -94,31 +116,3 @@ exports.handleEvent = async () => {
     csvContent
   );
 };
-
-function validateEnvironmentVariables() {
-  const requiredEnvVars = [
-    'BFF_BUCKET_NAME',
-    'BFF_BUCKET_PREFIX',
-    'WEB_LANDING_BUCKET_NAME',
-    'WEB_LANDING_BUCKET_PREFIX',
-    'FILE_NAME',
-    'CSV_CONFIGURATION_PARAMETER',
-    'GENERATE_INTERVAL',
-    'RADD_STORE_GENERATION_CONFIG_PARAMETER',
-    'RADD_STORE_REGISTRY_API_URL',
-    'SUBREGION_THRESHOLD',
-    'LOCALITY_THRESHOLD',
-    'POSTAL_CODE_THRESHOLD',
-    'ADDRESS_NUMBER_THRESHOLD',
-    'OVERALL_THRESHOLD',
-  ];
-
-  requiredEnvVars.forEach((envVar) => {
-    if (!process.env[envVar]) {
-      console.error(`Missing required environment variable: ${envVar}`);
-      throw new Error(`Missing required environment variable: ${envVar}`);
-    } else {
-      console.log(`Environment variable ${envVar} is set`);
-    }
-  });
-}
