+++ functions/raddStoreRegistryLambda/src/app/ssmParameter.js
@@ -41,7 +41,36 @@ async function getParameterFromLayer(parameterName) {
   }
 }
 
+function validateEnvironmentVariables() {
+  const requiredEnvVars = [
+    'BFF_BUCKET_NAME',
+    'BFF_BUCKET_PREFIX',
+    'WEB_LANDING_BUCKET_NAME',
+    'WEB_LANDING_BUCKET_PREFIX',
+    'FILE_NAME',
+    'CSV_CONFIGURATION_PARAMETER',
+    'GENERATE_INTERVAL',
+    'RADD_STORE_GENERATION_CONFIG_PARAMETER',
+    'RADD_STORE_REGISTRY_API_URL',
+    'SUBREGION_THRESHOLD',
+    'LOCALITY_THRESHOLD',
+    'POSTAL_CODE_THRESHOLD',
+    'ADDRESS_NUMBER_THRESHOLD',
+    'OVERALL_THRESHOLD',
+  ];
+
+  requiredEnvVars.forEach((envVar) => {
+    if (!process.env[envVar]) {
+      console.error(`Missing required environment variable: ${envVar}`);
+      throw new Error(`Missing required environment variable: ${envVar}`);
+    } else {
+      console.log(`Environment variable ${envVar} is set`);
+    }
+  });
+}
+
 module.exports = {
   retrieveGenerationConfigParameter,
   retrieveCsvConfiguration,
+  validateEnvironmentVariables,
 };
